Aggiunti fragment per header,footer. Aggiunti fogli di stile per login,footer, header(fare responsive). Aggiunta vista amministratore per gestire il catalogo. Completare GuestDAO e GuestBean per permettere l'acquisto anche agli utenti non registrati

+++ StorageSession/src/it/unisa/model/DetailsDAO.java
@@ -98,7 +98,7 @@ public class DetailsDAO {
 				      preparedStatement.setInt(1, product.getCode());
 				      preparedStatement.setInt(2, ordine);
 				      preparedStatement.setInt(3, product.getCartQuantity());
-				      preparedStatement.setInt(4, product.getPrice());
+				      preparedStatement.setDouble(4, product.getPrice());
 				      if(preparedStatement.executeUpdate()!=1) {
 				    	  return false;
 				      }
